Move prospector to its own package
