+++ filebeat/prospector/prospector_log.go
@@ -1,4 +1,4 @@
-package crawler
+package prospector
 
 import (
 	"os"
