+++ filebeat/prospector/prospector_test.go
@@ -1,25 +1,13 @@
 // +build !integration
 
-package crawler
+package prospector
 
 import (
 	"regexp"
 	"testing"
 
-	"github.com/elastic/beats/filebeat/input"
-	"github.com/elastic/beats/libbeat/common"
 	"github.com/stretchr/testify/assert"
 )
-
-func TestProspectorDefaultConfigs(t *testing.T) {
-
-	prospector, err := NewProspector(common.NewConfig(), *input.NewStates(), nil)
-	assert.NoError(t, err)
-
-	// Default values expected
-	assert.Equal(t, DefaultIgnoreOlder, prospector.config.IgnoreOlder)
-	assert.Equal(t, DefaultScanFrequency, prospector.config.ScanFrequency)
-}
 
 func TestProspectorInitInputTypeLogError(t *testing.T) {
